Video stream (#3)
* Iniital requirements for python environment

* Initial camera class and web server

This allows you to connect to a camera and acquire images from
it. The camera is also viewable from a Flask server on port 5000.

* Remove unnecessary function.

* Create PyGame GUI for PiTFT

This adds a class for displaying images to the PiTFT 2.4''
display. The images can now be seen both through Flask and
the display!

* Moved Flask app to library and made optional.

The Flask app seems to be causing some trouble anyway, so I decided
to make it optional. Everything ostensibly works, it just has
some ugly output whenever you quit it while its running.

The signal handlers were also getting in the way of each other
between the Flask server and the one I installed. Now they're
mutually exclusive.

* Templates hvae been moved to flask_app.

Just doing some cleanup.